+++ src/app/components/appFlow.tsx
@@ -1,0 +1,77 @@
+'use client';
+
+import { useState, useEffect } from 'react';
+import { useRouter } from 'next/navigation';
+import Image from 'next/image';
+import Onboarding from './onBoarding/onboarding';
+import LoginForm from './Auth/login-form';
+
+const isAuthenticated = () => {
+  return typeof window !== 'undefined' && localStorage.getItem('isAuthenticated') === 'true';
+};
+
+export default function AppFlow() {
+  const [showSplash, setShowSplash] = useState(true);
+  const [opacity, setOpacity] = useState(0);
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
+  const [showOnboarding, setShowOnboarding] = useState(false);
+  const router = useRouter();
+
+  useEffect(() => {
+    const fadeIn = setTimeout(() => {
+      setOpacity(100);
+    }, 100);
+
+    const showLoginTimeout = setTimeout(() => {
+      setShowSplash(false);
+      if (isAuthenticated()) {
+        setIsLoggedIn(true);
+        setShowOnboarding(true); // Mostrar el onboarding siempre después de login
+      }
+    }, 3000); // Reducir el tiempo de splash a 3 segundos para prueba
+
+    return () => {
+      clearTimeout(fadeIn);
+      clearTimeout(showLoginTimeout);
+    };
+  }, [router]);
+
+  const handleLogin = () => {
+    setIsLoggedIn(true);
+    localStorage.setItem('isAuthenticated', 'true');
+    setShowOnboarding(true); // Mostrar siempre el onboarding después de login
+  };
+
+  const handleOnboardingComplete = () => {
+    router.push('/home'); // Ir a home solo cuando el onboarding se complete
+  };
+
+  if (showSplash) {
+    return (
+      <div className="flex items-center justify-center h-screen bg-yellow-100">
+        <div
+          className={`transition-opacity duration-1000 ease-in`}
+          style={{ opacity: opacity / 100 }}
+        >
+          <Image
+            src="/altour.png"
+            alt="Altour Logo"
+            width={300}
+            height={300}
+            priority
+          />
+        </div>
+      </div>
+    );
+  }
+
+  if (!isLoggedIn) {
+    return <LoginForm onLogin={handleLogin} />;
+  }
+
+  if (showOnboarding) {
+    return <Onboarding onComplete={handleOnboardingComplete} />;
+  }
+
+  return null;
+}
